Decompose review view model
